Removed redux-simple-router, Updated git repo URLs
Both react-router-redux and redux-simple-router were listed in the dependencies. I removed redux-simple-router to clean it up since it has been renamed to the other module, and I also updated the "ericelliott/isomorphic-express-boilerplate" references to "cloverfield-tools/universal-react-boilerplate" just to keep it current despite the old URL redirecting to the new one.